requirements/base.txt
Django==1.11.6
django-environ==0.4.4
djangorestframework==3.7.7
djangorestframework-gis==0.12
dj-database-url==0.4.2
bleach==2.1.2
gunicorn==19.7.1
psycopg2==2.7.3.1
pytz==2017.2
whitenoise==3.3.1
